Import of OLIVER/CatalystX-ListFramework-Builder-0.19 from CPAN.
gitpan-cpan-distribution: CatalystX-ListFramework-Builder
gitpan-cpan-version:      0.19
gitpan-cpan-path:         OLIVER/CatalystX-ListFramework-Builder-0.19.tar.gz
gitpan-cpan-author:       OLIVER
gitpan-cpan-maturity:     released

+++ lib/CatalystX/ListFramework/Builder/Controller/Root.pm
@@ -26,8 +26,7 @@ sub default :Private {
 
 sub end :Private {
     my ($self, $c) = @_;
-    return if $c->res->output or ($c->res->status == 304);
-    $c->detach('JSON');
+    $c->detach('JSON') if $c->stash->{json_data};
 }
 
 sub err_message :Private {
